Fix `generateEServiceTemplateRiskAnalysisAnswersSQL` position
